Added level reward system
Leveling up now rewards you with a box opening mechanic, with three random options as rewards. One upgrades speed, one for gun damage and fire rate, and another for health upgrades. Added systems for updating these values and basic balancing implemented for the values. Should hopefully be kind of actually a game that is almost sort of fun!